Modify and create test classes for GestioneNewsController.java and GestioneNewsServiceImpl.java Modify type id  in GestioneNewsService.java,GestioneNewsDao.java,GestioneNewServiceImpl.java and add annotation in News.java

+++ hardship-api/src/main/java/io/hardship/hardshipapi/service/GestioneNewsService.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 public interface GestioneNewsService {
 
-    Optional<News> getNews(Long id);
+    Optional<News> getNews(Integer id);
     List<News> getAllNews();
     Optional<News> createNews(News news);
 }
